Branching respects the sign of the relative address

+++ lib/instruction/bpl.c
@@ -1,7 +1,8 @@
+#include <stdint.h>
 #include "bpl.h"
 
 void bpl(struct cpu *cpu) {
 	if (!cpu->negative) {
-		cpu->program_counter += cpu->operand;
+		cpu->program_counter += (int8_t)cpu->operand;
 	}
 }
